server-backend/requirements.txt
fastapi>=0.100.1
uvicorn>=0.23.0
pydantic~=2.8.2
starlette~=0.37.2
python-jose~=3.3.0
httpx~=0.27.0
httpx_auth~=0.22.0
bcrypt~=4.1.3
